--- src/components/BlogEditor.css
@@ -1,3 +1,0 @@
-.rdw-editor-main{
-    height: 400px;
-}
