+++ src/components/BlogEditor/index.css
@@ -1,0 +1,3 @@
+.rdw-editor-main{
+    height: 400px;
+}
