+++ src/pages/SandBox/BlogPreview/index.tsx
@@ -1,7 +1,7 @@
 import { PageHeader } from 'antd';
 import React, { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom';
-import { blogPreview } from '../../../services/SandBox/BlogPreview/BlogPreview'
+import { blogPreview } from '../../../services/SandBox/BlogPreview'
 
 const BlogPreview: React.FC = () => {
     const id = useParams().id!
