+++ src/pages/SandBox/Home/index.tsx
@@ -1,7 +1,7 @@
 import {  Card, List } from 'antd'
 import { HeartTwoTone } from '@ant-design/icons'
 import React, { useEffect, useState } from 'react'
-import { getAllBlog } from '../../../services/SandBox/Home/Home';
+import { getAllBlog } from '../../../services/SandBox/Home';
 
 const Home: React.FC = () => {
   const [blogList, setBlogList] = useState()
