--- src/services/Register/Register.tsx
@@ -1,8 +1,0 @@
-import axios from "axios"
-
-export const regiserSend = (values: Register.values) =>
-    axios.post('/accountManage/register', {
-        username: values.username,
-        password: values.password,
-        gender: values.gender
-    })
